HSoM MUI update
Updated some of the text and all of the examples in HSoM.MUI.  The
examples are now fully up to date with the new MUI system; however, the
text still needs some work.  Notably, this chapter should be after
arrows are introduced.

In the process, I created some helper functions (like accum, unique,
etc.) and put them in Control.SF.AuxFunctions.