Finished HSoM test cases
I encoded every test case that was suggested in HSoM into Euterpea.

Turns out we've got a pretty substantial bug in perf.

All of these test cases were supposed to be true, and anything that was
supposed to be true up to musical equivalence (as defined in the book)
turns up false in at least one test case.

The python script takes care of generating a test program for all the
tests in EuterpeaTests.hs (windows only)