Testing updates and bug fixes
I improved the automatic testing script so it saves errors and displays
more concise and better-presented console output.  I fixed the invert
bug and modified perf so that all of the tests pass when we restrict
ourselves to the default player. The fancy player, it seems, does not
handle bizarre input well.

See the comments in EuterpeaTests.hs and EuterpeaInstances.hs for
details. See error.log for specific failing cases.

*** Note: the testing script relies on quickcheck and ansi-terminal, so
install those with cabal before running.