Restricted haskell-src-exts version
The haskell-src-exts package was recently updated (to 1.14.0) and is no
longer backwards compatible (causing problems with ArrowWrap.hs).  I'm
not really sure what ArrowWrap does or why we need it, so a long term
fix would be to figure that out and update appropriately, but for now,
I've simply restricted the version to the last working one (1.13.5) in
the cabal file.