Issue #11 Fix - Removed Datatype Context
In Euterpea.IO.Audio.Types, there is an unused section labeled
"Experimental stuff" that defines an Unlifted class and some instances.
Within this section is a data type definition that uses a datatype
context that is giving some users problems.  The whole section has been
commented out.  (The context in question was also changed to a GADT
style in case this section is enabled in the future.)